requirements.txt
pandas
torchvision
pillow
torch